api/requirements.minimal.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Task queue
celery==5.3.4
kombu==5.3.4

# LangChain / LangGraph
langchain==0.2.14
langgraph==0.0.56

# Storage
minio==7.2.0
boto3==1.34.0

# Data processing
pandas==2.1.4
numpy==1.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# Excel processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# PDF processing
pdfplumber==0.10.3
PyPDF2==3.0.1
camelot-py[cv]==0.10.1
PyMuPDF==1.23.8
Pillow==10.1.0

# Data normalization
phonenumbers==8.13.26
usaddress==0.5.10
dateparser==1.2.0
python-dateutil==2.8.2

# NLP
spacy==3.7.2

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# HTTP client
httpx==0.25.2
tenacity==8.2.3

# Utilities
python-multipart==0.0.6
orjson==3.9.15

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
